Restricting user update on save to Craft Pro only
The `updateFoxyCartCustomer()` function was being called on the `onSaveUser` event, but it's functionality that requires Craft Pro. As such, for anyone not on Craft Pro it was triggering a permissions error when trying to save any user. This change ensures that that only happens if the Craft installation has the Pro edition.